pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ascii-tool"  # Unique name on PyPI
version = "0.1.3"
description = "A simple image to ascii converter"
readme = "README.md"
authors = [
    {name = "lukaseepydev", email = "lukaseepydev@gmail.com"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
dependencies = [
    "logging",
    "pillow",
    "argparse",
    "pyside6"
]

[tool.setuptools.packages.find]
where = ["src"]   # or ["ascii_tool"] if no src folder